services/pyproject.toml
[tool.poetry]
name = "ap2-box-platform-services"
version = "0.1.0"
description = "Microservices for the AP2 Box Platform"
authors = ["Team"]
packages = [{ include = "schema_registry" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
pydantic = "^1.10.0"
redis = "^5.0.1"
aioboto3 = "^12.0.0"
PyJWT = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
fakeredis = "^2.23.0"
moto = { version = "^5.0.0", extras = ["s3"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
